cleanup test scripts and setup codecov
